requirements.txt
dawgz==0.4.1
jupyter>=1.0.0
jupyterlab>=3.5.3
lampe>=0.8.1
matplotlib>=3.7.1
numpy>=1.22.0
PyYAML>=5.4.1
scipy>=1.10.1
torch>=1.13.0
tqdm==4.64.1
